+++ MoltenVK/MoltenVK/API/vk_mvk_moltenvk.h
@@ -557,14 +557,15 @@ typedef struct {
 	/**
 	 * Use MTLFence, if it is available on the device, for VkSemaphore synchronization behaviour.
 	 *
-	 * This parameter interacts with semaphoreUseMTLEvent. If both are enabled, semaphoreUseMTLFence
-	 * takes priority and MTLFence will be used if it is available, otherwise MTLEvent will be used
-	 * if it is available. If neither semaphoreUseMTLFence or semaphoreUseMTLEvent are enabled, or
-	 * if neither MTLFence or MTLEvent are available, CPU-based synchoronization will be used.
+	 * This parameter interacts with semaphoreUseMTLEvent. If both are enabled, on GPUs other than
+	 * NVIDIA, semaphoreUseMTLEvent takes priority and MTLEvent will be used if it is available,
+	 * otherwise MTLFence will be used if it is available. On NVIDIA GPUs, the opposite priority
+	 * applies. If neither semaphoreUseMTLFence nor semaphoreUseMTLEvent are enabled, or if neither
+	 * MTLEvent nor MTLFence are available, CPU-based synchoronization will be used.
 	 *
 	 * In the special case of VK_SEMAPHORE_TYPE_TIMELINE semaphores, MoltenVK will always
 	 * use MTLSharedEvent if it is available on the platform, regardless of the values of
-	 * MVK_ALLOW_METAL_FENCES or MVK_ALLOW_METAL_EVENTS.
+	 * semaphoreUseMTLEvent or semaphoreUseMTLFence.
 	 *
 	 * The value of this parameter must be changed before creating a VkDevice,
 	 * for the change to take effect.
@@ -573,21 +574,22 @@ typedef struct {
 	 * MVK_ALLOW_METAL_FENCES
 	 * runtime environment variable or MoltenVK compile-time build setting.
 	 * If neither is set, this setting is enabled by default, and VkSemaphore will use MTLFence,
-	 * if it is available.
+	 * if it is available, unless MTLEvent is available and semaphoreUseMTLEvent is enabled.
 	 */
 	VkBool32 semaphoreUseMTLFence;
 
 	/**
 	 * Use MTLEvent, if it is available on the device, for VkSemaphore synchronization behaviour.
 	 *
-	 * This parameter interacts with semaphoreUseMTLFence. If both are enabled, semaphoreUseMTLFence
-	 * takes priority and MTLFence will be used if it is available, otherwise MTLEvent will be used
-	 * if it is available. If neither semaphoreUseMTLFence or semaphoreUseMTLEvent are enabled, or
-	 * if neither MTLFence or MTLEvent are available, CPU-based synchoronization will be used.
+	 * This parameter interacts with semaphoreUseMTLFence. If both are enabled, on GPUs other than
+	 * NVIDIA, semaphoreUseMTLEvent takes priority and MTLEvent will be used if it is available,
+	 * otherwise MTLFence will be used if it is available. On NVIDIA GPUs, the opposite priority
+	 * applies. If neither semaphoreUseMTLFence nor semaphoreUseMTLEvent are enabled, or if neither
+	 * MTLEvent nor MTLFence are available, CPU-based synchoronization will be used.
 	 *
 	 * In the special case of VK_SEMAPHORE_TYPE_TIMELINE semaphores, MoltenVK will always
 	 * use MTLSharedEvent if it is available on the platform, regardless of the values of
-	 * MVK_ALLOW_METAL_FENCES or MVK_ALLOW_METAL_EVENTS.
+	 * semaphoreUseMTLEvent or semaphoreUseMTLFence.
 	 *
 	 * The value of this parameter must be changed before creating a VkDevice,
 	 * for the change to take effect.
@@ -596,7 +598,7 @@ typedef struct {
 	 * MVK_ALLOW_METAL_EVENTS
 	 * runtime environment variable or MoltenVK compile-time build setting.
 	 * If neither is set, this setting is enabled by default, and VkSemaphore will use MTLEvent,
-	 * if it is available, unless if MTLFence is available and semaphoreUseMTLFence is enabled.
+	 * if it is available.
 	 */
 	VkBool32 semaphoreUseMTLEvent;